# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GlutenFree documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/post-categories/store": {
            "post": {
                "tags": [
                    "Admin-Posts-Categories"
                ],
                "summary": "Create posts category",
                "description": "Store posts category",
                "operationId": "store-post-category",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/post-categories/update/{slug}": {
            "post": {
                "tags": [
                    "Admin-Posts-Categories"
                ],
                "summary": "Update posts category",
                "description": "Display a listing of the resource.",
                "operationId": "update-post",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Posts category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "aktualnosci"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/post-categories/delete/{slug}": {
            "post": {
                "tags": [
                    "Admin-Posts-Categories"
                ],
                "summary": "Delete posts categories by slug",
                "description": "Delete posts category",
                "operationId": "delete-posts-category",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Posts category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "aktualnosci"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Item not founded",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Item not founded"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/posts/store": {
            "post": {
                "tags": [
                    "Admin-Posts"
                ],
                "summary": "Create post",
                "description": "!!! use form-data for api !!!",
                "operationId": "store-post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Posts category language dont match with language of the post",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Posts category language dont match with language of the post"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/posts/update/{slug}": {
            "post": {
                "tags": [
                    "Admin-Posts"
                ],
                "summary": "Update post",
                "description": "!!! use form-data for api !!!",
                "operationId": "update-post",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Posts slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "co-to-jest-gluten"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "The provided slug is incorrect",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Item not founded"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePostRequest": {
                "title": "StorePostRequest",
                "description": "Class StorePostRequest",
                "required": [
                    "category_id",
                    "lang",
                    "title",
                    "h1",
                    "content",
                    "image"
                ],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "Posts category ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "lang": {
                        "title": "lang",
                        "description": "language of post",
                        "type": "string",
                        "example": "pl"
                    },
                    "title": {
                        "title": "title",
                        "description": "title",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "post post post"
                    },
                    "h1": {
                        "title": "h1",
                        "description": "h1",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "post post post"
                    },
                    "content": {
                        "title": "content",
                        "description": "content",
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 10,
                        "example": "this is my content wow"
                    },
                    "image": {
                        "title": "image",
                        "description": "image",
                        "type": "string",
                        "example": "Image file"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 10,
                        "example": "this is my description"
                    },
                    "keywords": {
                        "title": "keywords",
                        "description": "keywords",
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 10,
                        "example": "test, foo, baz"
                    },
                    "is_on_main": {
                        "title": "is_on_main",
                        "description": "is_on_main",
                        "type": "boolean",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UpdatePostRequest": {
                "title": "UpdatePostRequest",
                "description": "Class UpdatePostRequest",
                "required": [
                    "category_id",
                    "lang",
                    "title",
                    "h1",
                    "content",
                    "is_on_main"
                ],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "integer",
                        "type": "integer",
                        "example": "1"
                    },
                    "lang": {
                        "title": "lang",
                        "description": "language",
                        "type": "string",
                        "example": "pl"
                    },
                    "title": {
                        "title": "title",
                        "description": "title",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "post post"
                    },
                    "h1": {
                        "title": "h1",
                        "description": "h1",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "post post"
                    },
                    "content": {
                        "title": "content",
                        "description": "content",
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 10,
                        "example": "this is my content"
                    },
                    "image": {
                        "title": "image",
                        "description": "image",
                        "type": "string",
                        "example": "Image file"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 10,
                        "example": "this is my description"
                    },
                    "keywords": {
                        "title": "keywords",
                        "description": "keywords",
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 10,
                        "example": "test, foo, baz"
                    },
                    "is_on_main": {
                        "title": "is_on_main",
                        "description": "is_on_main",
                        "type": "boolean",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "StorePostCategoryRequest": {
                "title": "StorePostCategoryRequest",
                "description": "Class StorePostCategoryRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of posts category",
                        "type": "string",
                        "example": "Aktualności"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer \"token\")",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}